Make the "Plankton" name obsolete
Since the Image Service is not going to be decoupled from Cyclades
we will stop calling it with a different name because it is
frustrating for people who are trying to understand the Synnefo
architecture and the interaction between different services.

We will keep all references to "plankton" inside the code, but we
remove all references inside the documentation. From now on
"Plankton" will be just called "The Image Service" and we will note
that it is a part of Cyclades. The same already happens with
"The Network Service", which doesn't have a special name since it's
part of Cyclades too.

This patch removes all Plankton references inside the documentation,
except from those inside `docs/plankton-api-guide.rst' which will be
removed with a separate patch.