pithos-backend: disable AstakosClient if no URL given
If the backend is initialized with astakos_url=None,
or AstakosClient cannot be imported,
a fake AstakosClient is instantiated which will raise
an AssertionError on any attempt to access it.

Read-only uses of pithos backend, as in pithcat,
should never have to access the astakosclient initialized.
The fake client is there to make an access raise an informative
error message instead of 'NoneType object has no attribute xxx'.