astakos: Record the service originating a resource
Add field service_origin in model Resource as well as in the
resource definitions, indicating the service where the resource
comes from. It is not a foreign key, in order to allow deleting
a component/service, without affecting registered resources and
quota.

API call /service_quotas uses this field to identify relevant
resources.