pithos: change reconcile-resources-pithos command
This command syncronizes the astakos quota with the pithos db.
However, if there was an account in the astakos database but there
was not the relative path in the pithos db (new pithos database)
the account quotas were not synced.
This commit fixes this.

Also, sync all the registed resources defined for the service
(the resource to be synced used to be hardcoded before)