cyclades: Convert network to floating IP pool
Add '--floating-ip-pool' option to network-modify management command, in
order to convert a network to a floating IP pool. Conversation of a
network to a floating IP pool requires the creation of the network to
all Ganeti backends. Also, a check is added to to ensure that the link
is available in the Ganeti backend.

Signed-off-by: Christos Stavrakakis <cstavr@grnet.gr>