ci: turn system_uuid option into a hardcoded one
User doesn't need to specify the UUID of the owner of
system images (it's usually hard to find it).

Instead save the uuids for these users for our production and
testing deployments inside a list (DEFAULT_SYSTEM_IMAGES_UUID) and
use that (uuids never change). If you need to specify another user
it's preferable to specify the image_id using the corresponding option.