cyclades: extend VM info with 'port_forwarding'
Extend servers info API response with 'SNF:port_forwarding' attribute,
describing port fowarding rules (DNAT) that are applied to vms. The
description of such rules is done via the new CYCLADES_PORT_FORWARDING
setting. This setting can be either a dictionary mapping to ports to
tuples of (host, port), or a callable object that must return such a
tuple.   The caller will pass to the callable the following positional
arguments, in the following order:
* server_id: The ID of the VM in the DB
* ip_address: The IPv4 address of the public VM NIC
* fqdn: The FQDN of the VM
* user: The UUID of the owner of the VM