cyclades: Remove firewall tag when removing NIC
Remove firewall tag from ganeti instance when disconnecting them from
networks. Until now, this was not needed because a vm could not be
disconnected from public networks. Since floating IPs, a vm may be
dynamically be connected and disconnected to public networks, and so a
hanging firewall tag may affect another NIC.