deploy: Performance tune for postgresql
Turn off fsync, synchronous_commit and full_pages_writer to speed up
postgresql. DB is not no longer crash-safe, but this is no needed for
tests.

Also, introduce 'testing_vm' setting to distinquish between the type of
deployment. In cases like the above, we want to be able to configure the
server in a different way, whether is deployed for development/testing
or for production deployment.