cyclades: Fix bug in handling of reboot type.
'type' attribute of reboot OS API call was directly mapped to the
'reboot_type' option of Ganeti reboot job. However these two are
semantically different. The 'SOFT' and 'HARD' reboot types of OS API are
mapped to 'shutdown_timeout' option of Ganeti jobs. However, setting
this option from Ganeti RAPI is not supported. Until implemented, we
will map both types to 'hard' ganeti reboot, and fallback to default
shutdown timeout of Ganeti (120s).

Also, modify UI to not set the 'HARD' reboot type and fallback to
default option of API which is 'SOFT'.