cyclades: Fix bug in snf-dispatcher logic
This commit reverts commits #07602322 and #30514662 which introduced a
bug in dispatcher's logic. These commit fixed the issue that, in a case
the Cyclades DB is unsynced with Ganeti, a resource that does not exist
in the Ganeti backend, it  will not be deleted from DB in case the
remove job fails.

This commits fixes this issue, by quering Ganeti each time a failed
OP_INSTANCE_REMOVE or OP_NETWORK_REMOVE is processed.