cyclades: Fix bug during server creation
Commit #32a0b85, modified server creation, by refetching the object from
DB after issuing commission to quotaholder. However, the new object is
not having the 'config_url' attributed, which is added to the vm object
by the 'vmapi'. This commit, fixes this by moving the signal dispatching
code after refetching the vm. Also, this commit makes more clear the
exception handling code, when the creation job can not be enqueued to
Ganeti.