cyclades: refactor server creation command
Split the server creation command in more steps in order to be
decorated with the 'server_command' function and follow the logic of all
other server commands. The main change is that the commission job is not
accepted when the VM is stored in DB, but when the OP_INSTANCE_CREATE
job finishes in Ganeti, no matter whether it succeeded or not. Finally,
this commit includes a transaction testcase that checks instance
creation.