cyclades: Handle errors during vm creation
Commit #f2080d1 mocked an OP_INSTANCE_REMOVE when enqueuing an
OP_INSTANCE_CREATE job to Ganeti failed. However, this may result in
releasing IPs that may get reserved in Ganeti, because we can never
really know if the job has reached Ganeti or not (e.g.  timeout).  For
this reason, we should *never* delete the VM. Instead we mark the VM as
being in error state.  The user must explicitly delete the server.  This
has the downside, that we return 202 with the resulting VM being in
ERROR state. However, this is the only way to be really secure.