wheezy: Remove setup_environ
From Django docs:
"""
This function temporarily modified sys.path in order to make the parent
“project” directory importable under the old flat startproject layout.
This function is now deprecated, as its path workarounds are no longer
needed with the new manage.py and default project layout.

This function was never documented or part of the public API, but it was
widely recommended for use in setting up a “Django environment” for a
user script. These uses should be replaced by setting the
DJANGO_SETTINGS_MODULE environment variable or using
django.conf.settings.configure().

"""