cyclades: Simplify issue_and_accept_commission
Simplify the 'issue_and_accept_commission' function that handled
commissions that should be auto-accepted. Also do not 'reject' the
serial, since the resource that corresponds to this serial has already
been committed to DB. Finally, refactor 'floating_ip' code that used this
function by removing the 'commit_manually' decorator, since it is not
needed.