cyclades: Refactor reconciliation of pools
Move reconciliation of IP pools from 'snf-manage reconcile-networks'
to 'snf-manage reconcile-pools'. Also change the way reconciliation
works: Instead of reconciling the IP pools in DB with the IP pools in
Ganeti, the new reconciliation just checks that:
* there is no IPv4 address that is used by more than one NICs
* the only reserved values in the pools are the ones that are currently
  used by NICs
The reason why reconciliation of pools with Ganeti has been removed is
that there are too many race conditions, and that it may result in
strange incosistencies. As far as external reservations is concerned,
the new reconciliation mechanism just checks that each IP that is
reserved in a Ganeti backend is also reserved in the Cyclades DB.