common: Do not encode key paths as string
Handle explicitly key paths as lists of keys. This allows the key
to be any python object, as is the case in python dicts.

Provide utility unpack() to convert from a string representation to
a list of keys. It assumes that a key is either a string or an integer
(if it looks like one).