cyclades: Major refactor on NIC model
Major refactor on db models concerning NetworkInterface models. A new
table, namely 'IPAddress' has been created to represent an IPAddress
(IPv4 or IPv6). Each IPAddress is related with a network and the subnet
from which it has been allocated. Also it contains 'floating_ip' attribute
to indicate whether it is a floating or static IP.

Also the dummy SecurityGroup model has been created, although it will
not currently be used.

The NetworkInterface model has been extended with the 'name',
'device_owner' and 'security_groups' attributes.

Signed-off-by: Marios Kogias <marioskogias@grnet.gr>