cyclades: Inspect qh NotFound response
Since a user can now specify the project where a resource is assigned,
user-provided data do reach the quotaholder. A NotFound response is
thus not necessarily an internal error; the user may just not be a
member of the input project. Query the quotaholder to find out whether
this is the case and, if so, raise BadRequest.