cyclades: Fix bug when deleting instances
If OP_INSTANCE_REMOVE runs before OP_INSTANCE_CREATE the job will fail.
Dispatcher will ask Ganeti if the instance exists, and if not it will
release it's IPv4 address. However the instance may not exist because it
has not being created yet. This commit fixes this, by making
OP_INSTANCE_REMOVE depend on the OP_INSTANCE_CREATE job if the job has
not completed.