cyclades: Create port with specified IPv4 Address
Extend POST /ports API method to respsect 'fixed_ips' attribute. If this
attribute contains a valid IPv4 address, the new port will get this
address instead of trying to allocate one from the available IP pools of
the network. Specifying an IPv6 address is not supported or more than
one fixed IPs is not supported.