cyclades: Fix double double creation of IP log
The port creation command called 'associate_port_with_machine' function
in order to associate the port with the machine and create an
IPAddressLog entry. However, this is already done by the connect
function which sends the job in the Ganeti backend. This resulted in
duplicate entries in the IP accounting mechanism.