cyclades: Create server with existing ports
Extend POST /servers API call to support creation of servers that will
be using existing ports. Networking of the instance is defined via the
'networks' attribute which can contain either network UUIDs or port
UUIDs. In case of network UUIDs, the user can also request for a
specific IPv4 address. For example:

{"network":  [{"uuid": 41, "fixed_ip": "192.168.2.0"}, {"port": 42}]}

This patch also changes the 'floating_ips' attribute to contain not
IPv4 addresses of floating IPs, but the UUIDs of the floating IPs that
the user want to connect to the instance.