cyclades: Fix handling of old IP pools
Before implementing the Neutron API, each network had one IP pool that
contained the whole subnet CIDR. This changed since now we have multiple
IP Pools(with base and offset attributes) per network and this IP pools
do not contain the network, broadcast and gateway addreeses. This commit
fixes an old migration to add 'base' and 'offset' attribute to all IP
pool objects, and adds an extra step to old and not-initialized pools in
order to externally reserve those addresses.