cyclades: Create 'cyclades_network' service
Create 'cyclades_network' service which is part of Cyclades, of 'network'
type and with 'network' prefix to hold all API endpoints that are
relative with networks. So, for example, the endpoint for networks will
be: /cyclades/network/v2.0/networks

Also move 'cyclades.floating_ip' and 'cyclades.network.private'
resources to 'cyclades_network' service.