cyclades: Create IPv6 Addresses from dispatcher
When ports are created from the API they do not have an IPv6 address.
The IPv6 address is only available after the MAC address of the
NIC(port) is known which is only available after the message from the
Ganeti backend has been received. However, dispatcher did not created
correctly the IPv6 addresses. This commit fixes this issue along with
some small code refactoring.