cyclades: Fix network-create management command
Fix when network is created in the Ganeti backends. The network must be
created in the Ganeti backends only after IPv4 or/and IPv6 subnet has
been created. This commit moves the code that created the network in the
Ganet backend out of 'network.create' function to a separate function
that is called at the end of command.

Also, this commit fixes some small issues and removes unnecessary
options.