pithos: Do not put xseg request before map retrieval
During map_retr_archipelago, we should not put the xseg request which fetched
the map data, before actually retrieving them.

The ctypes "from_address" method, does not copy the data where the pointer
points. It simply creates a ctypes instance using the given memory. Ctypes
string_at method on the other hand, creates a new string object from the given
address, effectively copying the data out of the shared segment.

So, we should create the hashmap list from the request buffer before putting the
request and returning it (along with the associated buffer) to the segment.