burnin: Change --quiet logging behavior
When --quiet flag is used, burnin will not produce a logging
file at all (neither will log to stdout). This flag is used
when we simple want to check if there are stale servers from
previous runs (we can do this based on burnin's exit status)
but we don't want to produce any log files.