cyclades: Add name to Ganeti NICs
Add 'name' attribute to newly created NICs. Currently, the name of the
NICs will not be used to refer to NICs, as it will be in Synnefo 0.15.
Names are added only to ease the migration to version 0.15. Also, this
commit updates snf-dispatcher to preserve the name of the NIC, if any,
when updating instance NICs.

Finally, the commit updates handling of firewall tags by adding a second
tag that contains the name of the NIC. The tags with the name of the
NICs will be kept in Synnefo 0.15, and the old tags with the index of
the NIC will be removed.