cyclades: Fix removal of stale NICs in BUILD state
Until now, NICs that did not exist in the Ganeti backend and were in
BUILD state were removed after a timeout (BUILDING_NIC_TIMEOUT). This
was dangerous, because if  an OP_INSTANCE_CONNECT job had been "stuck"
in Ganeti, the NIC and it's IP address would be released. This commit
fixes this by removing such NICs only if the OP_INSTANCE_CONNECT has
been finished and the NIC does not exist in the Ganeti backend. Because,
querying Ganeti has a time overhead, this is only done for NICs that
have not been created too recently.