cyclades: Map reboot type to shutdown timeout
Fix handling of reboot type option. If 'hard' reboot type is specified
this is mapped to zero shutdown timeout of OP_INSTANCE_REBOOT Ganeti
job. This is only supported from Ganeti>=2.10 and snf-ganeti. In other
versions this options will be ignored, and we will fall back to Ganeti's
default shutdown timeout (120s).