astakos: Add project default in resources
Add field `project_default' in resources. This will serve as a skeleton to
fill project quota (both project-level and member-level limits) for
resources not specified by the application.

The field is initialized by consulting the default base quota (field
`uplimit'): It is set to inf if uplimit is inf, otherwise it is set to
zero.