cyclades: Fix wrong state of deleted networks
Before deleting a network, the network is marked as drained to be
excluded from automatic IP allocation. The state of the network that is
exposed in the API is always 'SNF:DRAINED' if the drained flag is set.
This resulted in deleted networks to appear as 'SNF:DRAINED'. This
commit fixes this issue by clearing the 'drained' flag when deleting the
network.

Also, this commit sets the 'drained' flag only when deleting public
networks that can take part in automatic IP allocation. For private
networs setting this flag is not needed.