burnin: If machine is ERROR we cannot delete ports
If the machine is not in ACTIVE state we cannot delete its ports.
This patch changes burnin so when cleaning up stale servers
(where there is the possibility that the machine is not ACTIVE)
it will not try to disconnect the server from the public network.
The port will be destroyed by it self and we will delete the
stale floating IP afterwards.