snf-burnin: Add new NIC after ping
Burnin has a race condition where if it tries to add a new NIC before
the server has been completely booted, the kernel will not detect the
new NIC (not until the server reboots).
Fix this by adding the new NIC to the server after we have tested (ping)
that the server is up and running.