deploy: Add Archipelago support
Extend snf-deploy to support installation and configuration of
Archipelago.

Archipelago related packages are found in apt.dev unstable an in
ceph's repo, which are added to the source.list.

A new nfs mount point is created /srv/archip and two new components,
Archip and ArchipGaneti are added to be included in pithos, cyclades,
ganeti and master roles.