cyclades: Fix quota bug in snf-dispatcher
Clear the 'vm.serial' field after processing a job that was not
expected. When processing such a job, we reject the pending commission
and issue a new one that reflects the unexpected job. However, since we
reject the commission that is associated with the 'vm.task_job_id' job,
we must also clear the expected serial. Otherwise, when the
'vm.task_job_id' completes, we will not issue a new commission as it is
needed.