plankton: Check that object is an image
Fix bug that caused Plankton to return Pithos files that were not
registered as images. Plankton did not checked that the requested Pithos
object (UUID or URL) was an Image. This was caused because
'get_object_meta' call of PithosBackend returns the common metadata for
all Pithos files (uuid, size, etc.) even if the file had no metadata in
the Plankton domain.