pithos: Lazy evaluation of "available" attribute
Whenever the backend tries to access the object's content
checks its database records whether the snapshot map is available or not.
If it is not, it pings Archipelago to check the status of the map.
If the map exists, then it updates the snapshot record in the Pithos
database and sets the "available" attribute to true.

Consequent Archipelago pings are limited.
The time interval limit for consequent Archipelago pings
is configured by the PITHOS_BACKEND_MAP_CHECK_INTERVAL setting.