deploy: workaround for java console Issue
Java VncViewer class has an issue with self-signed certificates.
With existing deployment it does not work if UI_MEDIA_URL uses
FQDN. Use the node's IP instead.

NOTE: In a linux (Debian) client add the following in
.java/deployment/deployment.properties:

deployment.security.level=MEDIUM

which will prompt to allow execution of the applet.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>