deploy: Change gunicorn file only if Archip is installed
Currently cyclades and pithos require a specific gunicorn setup
with GeventArchipelagoWorker. Plus they require a config
file that is included in snf-pithos-app; pithos.conf.py

This file must be installed both in pithos and cyclades. Therefore
we have it as a template and install it under /etc/archipelago.

Add archipelago.conf template in order to customize SEGMENT_SIZE.

For the above responsible is Archip component.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>