snf-ci: Fix the branch name for pithos webclient
When cloning the pithos webclient, snf-ci tries to checkout a branch
with the same name as the synnefo one (if such branch exists). To do so
it gets the list of all branches using the 'git branch' command. But
'git branch' uses the pager and so the command's output is filled with
garbage. This patch fixes that by telling git to not use the pager.