cyclades: Create common '-u/--user' option
Create 'cli_option' module, to contain the common definitions for
options of Synnefo management commands.

Currently, it contains only the '--user' option definition which accepts
user's UUID or email and annotates the command options with the
'user_id' and 'user_email' options.

Finally, the module contains a helper function for creating boolean
options that will parse the option value(str) to a boolean value.