ci: Add ssh port option
Add an option to connect to a specific ssh port of the created VM.
It is mainly useful when running ci from a VM in demo.synnefo.org. In
this case, we don't want to deduct the ssh port from the returned
server IP since the demo's DNAT applies only to connections out of the
demo's private network.