Implement NewlineStreamHandler
When StreamHandler writes a formatted log message to its stream, it adds
a newline terminator. This behavior is inherited by FileHandler and the
other classes which derive from it (such as the rotating file handlers).

Starting with Python 3.2, the message terminator will be configurable.
This has been done by adding a terminator attribute to StreamHandler,
which when emitting an event now writes the formatted message to its
stream first, and then writes the terminator. If you don't want newline
termination for a handler, just set the handler instance's terminator
attribute to the empty string.

This class implements python's 3.2 StreamHandler.

Ref #3: Log all stdout/stderr for snf-manage invocations