snf-ci: Don't get the schema files from local repo
snf-ci used to find the schema files in the local repo and upload them
to the testing server. This is wrong since one can choose to test a
different branch that the one he has currently checked out, so snf-ci
will use the wrong schema files.

This patch fixes the above problem by instructing snf-ci to use the
schema files it finds in the repo cloned inside the testing server.