burnin: Fix 'project_id' param in 'create_floatingip'
Until now, burnin was using a 'feature' version of kamaki
where the 'create_floatingip' method had a 'project' parameter.
This parameter was renamed to 'project_id' for uniformity with the
other kamaki methods.

These changes have been merged to 'develop' branch of kamaki and the API
is not expected to be changed again. This patch changes the way burnin
invokes 'create_floatingip' to use the 'project_id' parameter instead of
the 'project' one.